+++ python/examples/spot_cam/webrtc.py
@@ -21,6 +21,9 @@ from aiortc import (
 from aiortc.contrib.media import MediaRecorder
 import requests
 
+import cv2
+import numpy as np
+
 from bosdyn.client.command_line import (Command, Subcommands)
 from webrtc_client import WebRTCClient
 
@@ -170,6 +173,10 @@ async def process_frame(client, options, shutdown_flag):
             frame = await client.video_frame_queue.get()
 
             if options.count == 0:
+                pil_image = frame.to_image()
+                cv_image = np.array(pil_image) 
+                cv2.imshow('display', cv_image)
+                cv2.waitKey(1)
                 continue
 
             frame.to_image().save(f'{options.dst_prefix}-{count}.jpg')